move i18n dictionaries
