+++ src/utils/i18nDictionaryFr.js
